refactor: move `validation` and `utils` to a shared directory and update imports
