+++ src/main/models/holidays.js
@@ -1,6 +1,7 @@
-import { assertValidCalendarDate } from 'utils/dateHelper.js'
-import { NotFoundError, ValidationError } from 'infra/errors.js'
-import booleanHelper from 'utils/booleanHelper.js'
+import { assertValidCalendarDate } from 'shared/utils/date.js'
+import { NotFoundError } from 'shared/errors.js'
+import booleanHelper from 'shared/utils/boolean.js'
+import { validateHoliday, validateHolidayUpdate } from 'shared/validators/holidays.js'
 
 function normalizeHoliday(row) {
   if (!row) return row
@@ -17,7 +18,7 @@ export default class Holidays {
   }
 
   create(holiday) {
-    this._validateHoliday(holiday)
+    validateHoliday(holiday)
 
     let { description, type, date, is_business_day, should_count_as_business_day } = holiday
 
@@ -68,7 +69,7 @@ export default class Holidays {
   }
 
   update(id, data) {
-    this._validateHolidayUpdate(data)
+    validateHolidayUpdate(data)
 
     const currentHoliday = this.getById(id)
 
@@ -123,92 +124,4 @@ export default class Holidays {
     return true
   }
 
-  _validateHoliday(data) {
-    this._assertDataObject(data)
-
-    assertValidCalendarDate(data.date, 'date')
-
-    const errors = []
-
-    if (!data.description || typeof data.description !== 'string') {
-      errors.push('description is required and must be a string')
-    }
-
-    if (typeof data.type !== 'string') {
-      errors.push('type must be a string')
-    }
-
-    if (typeof data.is_business_day !== 'boolean') {
-      errors.push('is_business_day must be a boolean')
-    }
-
-    if (typeof data.should_count_as_business_day !== 'boolean') {
-      errors.push('should_count_as_business_day must be a boolean')
-    }
-
-    if (errors.length > 0) {
-      throw new ValidationError({
-        message: 'invalid data was provided',
-        cause: errors
-      })
-    }
-  }
-
-  _validateHolidayUpdate(data) {
-    this._assertDataObject(data)
-
-    const errors = []
-    const allowedFields = [
-      'description',
-      'type',
-      'date',
-      'is_business_day',
-      'should_count_as_business_day'
-    ]
-
-    Object.keys(data).forEach((key) => {
-      if (!allowedFields.includes(key)) {
-        errors.push(`${key} is not a valid field`)
-      }
-    })
-
-    if (data.description !== undefined && typeof data.description !== 'string') {
-      errors.push('description must be a string')
-    }
-
-    if (data.date !== undefined) {
-      assertValidCalendarDate(data.date, 'date')
-    }
-
-    if (data.type !== undefined && typeof data.type !== 'string') {
-      errors.push('type must be a string')
-    }
-
-    if (data.is_business_day !== undefined && typeof data.is_business_day !== 'boolean') {
-      errors.push('is_business_day must be a boolean')
-    }
-
-    if (
-      data.should_count_as_business_day !== undefined &&
-      typeof data.should_count_as_business_day !== 'boolean'
-    ) {
-      errors.push('should_count_as_business_day must be a boolean')
-    }
-
-    if (errors.length > 0) {
-      throw new ValidationError({
-        message: 'invalid data was provided',
-        cause: errors
-      })
-    }
-  }
-
-  _assertDataObject(data) {
-    if (!data || typeof data !== 'object' || Array.isArray(data)) {
-      throw new ValidationError({
-        message: 'invalid data was provided',
-        cause: ['a holiday must be provided']
-      })
-    }
-  }
 }
