+++ src/renderer/src/components/payables/EditPayableModal.jsx
@@ -2,6 +2,7 @@ import { useEffect, useState } from 'react'
 import { toast } from 'sonner'
 
 import { centsToDecimalString, decimalToCents } from '@/lib/utils'
+import { validatePayableUpdatePayload } from 'shared/validators/payables.js'
 
 const emptyForm = {
   history: '',
@@ -74,33 +75,32 @@ function EditPayableModal({ open, payable, onClose, onSave }) {
   async function handleSubmit(event) {
     event.preventDefault()
 
-    if (!form.history.trim()) {
-      toast.error('History is required.')
-      return
-    }
-
-    if (!form.due_date) {
-      toast.error('Due date is required.')
-      return
-    }
-
     const centsValue = decimalToCents(form.value)
     if (centsValue === null) {
       toast.error('Value must be a valid amount with up to 2 decimal places.')
       return
     }
 
+    const payload = {
+      history: form.history.trim(),
+      value: centsValue,
+      due_date: form.due_date,
+      preferred_date: form.preferred_date || null,
+      invoice_id: form.invoice_id.trim() || null,
+      account_id: form.account_id.trim() || null
+    }
+
+    try {
+      validatePayableUpdatePayload(payload)
+    } catch (error) {
+      toast.error(Array.isArray(error.cause) ? error.cause[0] : error.message)
+      return
+    }
+
     setIsSaving(true)
 
     try {
-      const hasSaved = await onSave(payable.id, {
-        history: form.history.trim(),
-        value: centsValue,
-        due_date: form.due_date,
-        preferred_date: form.preferred_date || null,
-        invoice_id: form.invoice_id.trim() || null,
-        account_id: form.account_id.trim() || null
-      })
+      const hasSaved = await onSave(payable.id, payload)
 
       if (hasSaved) {
         onClose()
