+++ src/renderer/src/components/recurring/EditRecurringModal.jsx
@@ -2,6 +2,7 @@ import { useEffect, useState } from 'react'
 import { toast } from 'sonner'
 
 import { centsToDecimalString, decimalToCents } from '@/lib/utils'
+import { validateRecurringUpdatePayload } from 'shared/validators/recurring.js'
 
 const emptyForm = {
   history: '',
@@ -70,32 +71,30 @@ function EditRecurringModal({ open, recurring, onClose, onSave }) {
   async function handleSubmit(event) {
     event.preventDefault()
 
-    if (!form.history.trim()) {
-      toast.error('Description is required.')
-      return
-    }
-
-    const dueDay = Number(form.due_day)
-    if (!Number.isInteger(dueDay) || dueDay < 1 || dueDay > 31) {
-      toast.error('Due day must be an integer between 1 and 31.')
-      return
-    }
-
     const centsValue = decimalToCents(form.value)
     if (centsValue === null) {
       toast.error('Value must be a valid amount with up to 2 decimal places.')
       return
     }
 
+    const payload = {
+      history: form.history.trim(),
+      value: centsValue,
+      due_day: Number(form.due_day),
+      should_postpone: Boolean(form.should_postpone)
+    }
+
+    try {
+      validateRecurringUpdatePayload(payload)
+    } catch (error) {
+      toast.error(Array.isArray(error.cause) ? error.cause[0] : error.message)
+      return
+    }
+
     setIsSaving(true)
 
     try {
-      const hasSaved = await onSave(recurring.id, {
-        history: form.history.trim(),
-        value: centsValue,
-        due_day: dueDay,
-        should_postpone: Boolean(form.should_postpone)
-      })
+      const hasSaved = await onSave(recurring.id, payload)
 
       if (hasSaved) {
         onClose()
