+++ src/shared/errors.js
