+++ vitest.config.mjs
@@ -13,7 +13,7 @@ export default defineConfig({
     coverage: {
       provider: 'v8',
       reporter: ['text', 'json', 'html'],
-      include: ['src/main/infra/**/*.js', 'src/main/models/**/*.js'],
+      include: ['src/main/infra/**/*.js', 'src/main/models/**/*.js', 'src/shared/**/*.js'],
       exclude: ['src/main/infra/migrations/**']
     }
   }
